fix: change URL for server-database connection
In Docker, container names act as the "localhost" for services within
the same network, allowing them to communicate with each other using
the container name as a hostname. This eliminates the need for IP
addresses, as Docker’s internal DNS resolves the container names to the
correct IP addresses, enabling seamless interaction between containers.